phase q1 qa fixes
