+++ crates/atm-core/Cargo.toml
@@ -5,7 +5,7 @@ edition.workspace = true
 rust-version.workspace = true
 authors.workspace = true
 license.workspace = true
-description = "Daemon-free core library for local agent team mail workflows."
+description = "Core library for ATM store, transport, and workflow boundaries."
 repository.workspace = true
 homepage.workspace = true
 documentation = "https://docs.rs/agent-team-mail-core"
